+++ packages/kit/src/views/Overlay/BaseMenu.tsx
@@ -1,0 +1,48 @@
+import type { FC, ReactElement } from 'react';
+import { Children, cloneElement } from 'react';
+
+import { useIntl } from 'react-intl';
+
+import type { ICON_NAMES } from '@onekeyhq/components';
+import { Menu } from '@onekeyhq/components';
+
+import type { IMenuProps } from 'native-base';
+import type { MessageDescriptor } from 'react-intl';
+
+type IMenuOptions = (
+  | {
+      id: MessageDescriptor['id'];
+      onPress: () => void;
+      icon: ICON_NAMES;
+    }
+  | false
+  | undefined
+)[];
+
+export type IMenu = Omit<IMenuProps, 'trigger'>;
+interface IBaseMenu extends IMenu {
+  options: IMenuOptions;
+}
+
+const BaseMenu: FC<IBaseMenu> = ({ options, children, ...rest }) => {
+  const intl = useIntl();
+  return (
+    <Menu
+      w={190}
+      trigger={(triggerProps) =>
+        cloneElement(Children.only(children as ReactElement), triggerProps)
+      }
+      {...rest}
+    >
+      {options.filter(Boolean).map(({ onPress, icon, id }) => (
+        <Menu.CustomItem icon={icon} onPress={onPress}>
+          {intl.formatMessage({
+            id,
+          })}
+        </Menu.CustomItem>
+      ))}
+    </Menu>
+  );
+};
+
+export default BaseMenu;
